refactor: Rename 'shared' folder to 'common'
- Renamed the 'shared' directory to 'common' to better align with the project's organizational conventions.
- Updated all relevant import paths to reflect the new directory name, ensuring consistent access to decorators, entities and guards.
- This change enhances clarity and follows common NestJS project structure practices, making the codebase more intuitive for future development and collaboration.
